// dm-backend/src/modules/grpc/proto/v2.pb.ts
export interface GetPenalizedOrdersResponse_PenalizedOrders {
  productName: string;
  orderNumber: string;
  payoutAmount: number;
  penalty: number;
  finalPayout: number;
  nctReason: string;
  nctReasonAR: string;
  dmoId: string;
}